新增访客数据

+++ RedHeart-System/src/main/java/com/sinothk/redheart/domain/UserAO.java
@@ -86,5 +86,5 @@ public class UserAO {
     private String imei;
 
     @ApiModelProperty(value = "距离")
-    private Double distance;
+    private Long distance;
 }
